solution

+++ 1899D.cpp
@@ -4,15 +4,15 @@ using namespace std;
 int main(){
 	int tc;cin>>tc;
 	while(tc--){
-		int n;cin>>n;
-		vector<int>arr(n);
-		map<int,int>mp;
-		for (int i = 0; i < n; i++)
+		long long n;cin>>n;
+		vector<long long>arr(n);
+		map<long long,long long>mp;
+		for (long long i = 0; i < n; i++)
 		{
 			cin>>arr[i];
 			mp[arr[i]]++;
 		}
-		int ans=0;
+		long long ans=0;
 		for (auto i: mp)
 		{
 			ans+=((i.second*(i.second-1))/2);   //number of pairs
